app/requirements.txt
azure-functions==1.21.3
requests==2.32.3